Print numbers

+++ 0x06-pointers_arrays_strings/101-print_number.c
@@ -9,7 +9,7 @@ void print_number(int n)
 {
 	unsigned int tmp;
 	
-	tmp = (unsigned int) n;
+	tmp = n;
 	if (n < 0)
 	{
 		_putchar('-');
